BugFixes please merge to release.
Distribute matrixmarket.hh
AMG: make x consistent even if there is only one level
galerkin.hh: Initialize Matrixblock correctly
splvers.hh: Always call pre mthod of the preconditioner _before_ using
the operator. Thus x is consistent and the computed defect will be,
too.
matrixmarket.hh: use int instead of size_t as template parameter as
this is the case for the FieldMatrix, too.

[[Imported from SVN: r1179]]

+++ dune/istl/matrixmarket.hh
@@ -81,7 +81,7 @@ namespace Dune
     template<class M>
     struct mm_header_printer;
 
-    template<typename T, typename A, size_t i, size_t j>
+    template<typename T, typename A, int i, int j>
     struct mm_header_printer<BCRSMatrix<FieldMatrix<T,i,j>,A> >
     {
       static void print(std::ostream& os)
@@ -91,7 +91,7 @@ namespace Dune
       }
     };
 
-    template<typename T, size_t i, size_t j>
+    template<typename T, int i, int j>
     struct mm_header_printer<FieldMatrix<T,i,j> >
     {
       static void print(std::ostream& os)
@@ -112,7 +112,7 @@ namespace Dune
     template<class M>
     struct mm_block_structure_header;
 
-    template<typename T, typename A, size_t i, size_t j>
+    template<typename T, typename A, int i, int j>
     struct mm_block_structure_header<BCRSMatrix<FieldMatrix<T,i,j>,A> >
     {
       typedef BCRSMatrix<FieldMatrix<T,i,j>,A> M;
@@ -124,7 +124,7 @@ namespace Dune
       }
     };
 
-    template<typename T, size_t i, size_t j>
+    template<typename T, int i, int j>
     struct mm_block_structure_header<FieldMatrix<T,i,j> >
     {
       typedef FieldMatrix<T,i,j> M;
